requirements.txt
# Core framework
streamlit==1.34.0

# OpenAI LLM API
openai==1.30.1

# For managing environment variables
python-dotenv==1.0.1

# For storing and processing user memory
pandas==2.2.2

# Optional: Text similarity for memory context injection
scikit-learn==1.4.2

# Optional: Emoji support or mood charts
emoji==2.11.0
matplotlib==3.8.4

# For JSON-based storage or file handling
diskcache==5.6.3

# Optional: For summarizing memory or GPT preprocessing
tiktoken==0.6.0

sqlalchemy==2.0.30
psycopg2-binary==2.9.9